# src/main/java/com/Spring/LearningHabits/FileUploadDownloadAPI/controller/DownloadFileApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.18).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.Spring.LearningHabits.FileUploadDownloadAPI.controller;

import com.Spring.LearningHabits.FileUploadDownloadAPI.models.ErrorResponse;
import org.springframework.core.io.Resource;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.net.MalformedURLException;
import java.util.List;
import java.util.Map;

@Api(value = "downloadFile", description = "the downloadFile API")
@Validated
public interface DownloadFileApi {


    @ApiOperation(value = "donload file that you uploaded", nickname = "downloadFile", notes = "This api will allow you to donload file.", response = Resource.class, tags={ "fileHandling", })
    @ApiResponses(value = {
        @ApiResponse(code = 200, message = "successful operation", response = Resource.class),
        @ApiResponse(code = 400, message = "Input Validation Error", response = ErrorResponse.class),
        @ApiResponse(code = 403, message = "If user is not allowed to upload file", response = ErrorResponse.class),
        @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorResponse.class) })
    @GetMapping(value = "/downloadFile",
        produces = { "application/pdf", "application/json" })

    ResponseEntity<Resource> downloadFile() throws MalformedURLException;

}