# src/main/java/com/Spring/LearningHabits/FileUploadDownloadAPI/controller/UploadFileApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.18).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.Spring.LearningHabits.FileUploadDownloadAPI.controller;

import com.Spring.LearningHabits.FileUploadDownloadAPI.models.ErrorResponse;
import com.Spring.LearningHabits.FileUploadDownloadAPI.models.FileDisplayDetails;
import org.springframework.core.io.Resource;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.io.IOException;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-09-18T18:26:37.935292100+05:30[Asia/Calcutta]")
@Api(value = "uploadFile", description = "the uploadFile API")
public interface UploadFileApi {

    @ApiOperation(value = "Add a new file", nickname = "uploadDocument", notes = "This APi will upload file", response = FileDisplayDetails.class, tags={ "fileHandling", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successful operation", response = FileDisplayDetails.class),
        @ApiResponse(code = 400, message = "Input Validation Error", response = ErrorResponse.class),
        @ApiResponse(code = 403, message = "If user is not allowed to upload file", response = ErrorResponse.class),
        @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorResponse.class) })
    @RequestMapping(value = "/uploadFile",
        produces = { "application/json" }, 
        consumes = { "multipart/form-data" },
        method = RequestMethod.POST)
    ResponseEntity<FileDisplayDetails> uploadDocument(@ApiParam(value = "file detail") @Valid @RequestPart("file") MultipartFile file
) throws IOException;

}